Move index impls into arcon crate + clean arcon_macros
