+++ Cargo.toml
@@ -27,7 +27,7 @@ exclude = [
 [features]
 default = ["serde"]
 arcon_serde = ["serde_json", "bincode", "serde", "uuid/serde"]
-arcon_arrow = ["arrow", "arcon_macros/arcon_arrow"]
+arcon_arrow = ["arrow", "datafusion", "arcon_macros/arcon_arrow"]
 unsafe_flight = ["abomonation", "abomonation_derive", "arcon_macros/unsafe_flight"]
 #kafka = ["rdkafka", "futures", "serde_json", "serde"]
 thread_pinning = ["kompact/thread_pinning"]
@@ -49,6 +49,7 @@ uuid = "0.8"
 cfg-if = "0.1.10"
 crossbeam-utils = "0.7"
 num_cpus = "1.0"
+hierarchical_hash_wheel_timer = "1.0"
 hocon = {version = "0.3", default-features = false, features = ["serde-support"]}
 
 # Hashing
@@ -73,6 +74,7 @@ rayon = { version = "1.3.0", optional = true }
 abomonation = { version = "0.7.3", optional = true }
 abomonation_derive = { version = "0.5.0", optional = true }
 arrow = { git = "https://github.com/Max-Meldrum/arrow.git", branch = "as_ref_patch", optional = true }
+datafusion = { git = "https://github.com/Max-Meldrum/arrow.git", branch = "as_ref_patch", optional = true }
 
 [dev-dependencies]
 tempfile = "3"
