--- arcon_state/arcon_state_derive/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "arcon_state_derive"
-version = "0.1.0"
-authors = ["Max Meldrum <mmeldrum@kth.se>"]
-edition = "2018"
-
-[lib]
-proc_macro = true
-
-[dependencies]
-proc-macro2 = "1"
-quote = "1"
-syn = "1"
