+++ benches/appender.rs
@@ -1,7 +1,8 @@
 // Copyright (c) 2020, KTH Royal Institute of Technology.
 // SPDX-License-Identifier: AGPL-3.0-only
 
-use arcon_state::{Backend, EagerAppender, *};
+use arcon::prelude::{AppenderIndex, Backend, EagerAppender};
+use arcon_state::{with_backend_type, BackendType};
 use criterion::{criterion_group, criterion_main, Bencher, BenchmarkId, Criterion};
 use std::sync::Arc;
 use tempfile::tempdir;
@@ -32,7 +33,6 @@ fn appender(c: &mut Criterion) {
         */
 
     let window_size = WINDOW_SIZE;
-    #[cfg(feature = "sled")]
     group.bench_with_input(
         BenchmarkId::new("Mean Index Eager Sled", ""),
         &(window_size),
