+++ src/index/hash_table/eager.rs
@@ -1,14 +1,14 @@
 // Copyright (c) 2020, KTH Royal Institute of Technology.
 // SPDX-License-Identifier: AGPL-3.0-only
 
-use crate::{
+use crate::index::IndexOps;
+use arcon_state::{
     backend::{
         handles::{ActiveHandle, BoxedIteratorOfResult, Handle},
         Backend, MapState,
     },
     data::{Key, Value},
     error::*,
-    index::IndexOps,
 };
 use std::sync::Arc;
